update category list
